doc: Add requirements file
It's a good pratice in python to add a file called "requirements.txt",
with it, once someone is trying to install and run our application,
he/she must just type in "pip install -r requirements.txt" and everityng
will be installed. I'm not shure if we need a specific version, so I'm
using the most recent one (no version specified = newest one).